requirements.txt
# Dépendances
Pillow>=9.0.0